api/requirements.txt
fastapi
unicorn
python-multipart
paddlepaddle
# paddlepaddle-gpu
scikit-image
imgaug
pyclipper
lmdb
tqdm
numpy
rapidfuzz
opencv-python<=4.6.0.66
opencv-contrib-python<=4.6.0.66
cython
Pillow>=10.0.0
pyyaml
requests
pymupdf
premailer